Dodanie nowych modeli

+++ Projekt/Models/Products.cs
@@ -11,8 +11,8 @@ namespace Projekt.Models
         public int Id { get; set; }
         public string Type { get; set; }
         public string Name { get; set; }
-        public float Quantity { get; set; }
-        public float Price { get; set; }
+        public double Quantity { get; set; }
+        public double Price { get; set; }
 
         }
     }
